+++ loader/mobfile.go
@@ -14,7 +14,7 @@ type mobFileEntry struct {
 	DescriptionInRoom   string         `json:"description_in_room"`
 	WanderingDefinition WanderingEntry `json:"wandering_definition"`
 	Flags               []string       `json:"flags"`
-	MaxHealth           int            `json:"max_health"`
+	MaxHealth           int64          `json:"max_health"`
 }
 
 type WanderingEntry struct {
